% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{df_obj}
\alias{df_obj}
\title{Evaluate Gradient of the Objective Function}
\usage{
df_obj(g, N, mu, sigma, L, moments, stepsize, w0 = 0.1)
}
\arguments{
\item{g}{The parameter \eqn{g}.}

\item{N}{The parameter \eqn{N}, it should be larger than 2.}

\item{mu}{The parameter \eqn{\mu}, mean of the normal distribution.}

\item{sigma}{The parameter \eqn{\sigma}, standard deviation of the normal
distribution.}

\item{L}{Length of the samples to generate.}

\item{moments}{A vector of the true sample moments, 1st moment,
2nd - 4th central moments.}

\item{stepsize}{A vector of four scalars, step size for each of the 
four parameters.}

\item{w0}{The parameter \eqn{w_0}, default to 0.1.}
}
\value{
A vector of four partial derivatives.
}
\description{
Evaluate gradient of the objective function by Finite Difference (FD).
}
\examples{
Y = sim_Y(0.2, 3.5, 0.5, 0.1, 1000)
moments = mom(Y)
stepsize = rep(0.01, 4)
df_obj(0.3, 3, 0.4, 0.05, 1000, moments, stepsize)

}
