% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{df_obj2}
\alias{df_obj2}
\title{Evaluate Single Partial Derivative}
\usage{
df_obj2(i, theta, L, moments, h = 0.01, G = 10L, w0 = 0.1)
}
\arguments{
\item{i}{An integer, 1,2,3,4 corresponding to \eqn{g, N, \mu, \sigma}
respectively.}

\item{theta}{A vector of the parameters \eqn{g, N, \mu, \sigma}.}

\item{L}{Length of the samples to generate.}

\item{moments}{A vector of the true sample moments, 1st moment,
2nd - 4th central moments.}

\item{h}{A scalar, step size for the parameter change.}

\item{G}{A integer, default to 10, number of replications.}

\item{w0}{The parameter \eqn{w_0}, default to 0.1.}
}
\value{
A scalar.
}
\description{
Evaluate single partial derivative of the objective function by
Finite Difference (FD).
}
\examples{
theta = c(0.3, 3, 0.4, 0.05)
Y = sim_Y(0.2, 3.5, 0.5, 0.1, 1000)
moments = mom(Y)
df_obj2(2, theta, 1000, moments)

}
