% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{df_obj_beta}
\alias{df_obj_beta}
\title{Evaluate Gradient of the Objective Function}
\usage{
df_obj_beta(g, N, a, b, L, moments, stepsize, w0 = 0.1)
}
\description{
Analog to [df_obj()], but here the model is set with beta distribution.
}
\seealso{
[df_obj()]
}
