% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f_obj_beta}
\alias{f_obj_beta}
\title{Evaluate the Objective Function}
\usage{
f_obj_beta(g, N, a, b, L, moments, w0 = 0.1)
}
\description{
Analog to [f_obj()], but here the model is set with beta distribution.
}
\seealso{
[f_obj()]
}
