% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intq}
\alias{intq}
\title{Evaluate the Integral Numerically}
\usage{
intq(g, N, mu, sigma, ub, lb, n = 10L, w0 = 0.1)
}
\arguments{
\item{g}{The parameter \eqn{g}.}

\item{N}{The parameter \eqn{N}, it should be larger than 2.}

\item{mu}{The parameter \eqn{\mu}, mean of the normal distribution.}

\item{sigma}{The parameter \eqn{\sigma}, standard deviation of the normal
distribution.}

\item{ub}{Upper bound of the integral, \eqn{\int_{lb}^{ub}q(x)dx}.}

\item{lb}{Lower bound of the integral.}

\item{n}{Segments to divide when computing the integral numerically.}

\item{w0}{The parameter \eqn{w_0}, default to 0.1.}
}
\value{
A scalar value.
}
\description{
Evaluate the integral with given upper and lower bounds by using 
trapezoidal rule.
}
\examples{
intq(0.2, 3.5, 0.5, 0.1, 0.4, 0)

}
\seealso{
[q()]
}
