% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intq_beta}
\alias{intq_beta}
\title{Evaluate the Integral Numerically}
\usage{
intq_beta(g, N, a, b, ub, lb, n = 10L, w0 = 0.1)
}
\description{
Analog to [intq()], but here the model is set with beta distribution.
}
\seealso{
[intq()]
}
