% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mom}
\alias{mom}
\title{Compute the first four (central) moments}
\usage{
mom(samples)
}
\arguments{
\item{samples}{A vector of the samples.}
}
\value{
A vector of the first four (central) moments. 
  \deqn{(E[x],E[(x-E[x])^2],E[(x-E[x])^3],E[(x-E[x])^4])}
}
\description{
`mom()` compute the first moment, the second to fourth central moments.
}
\examples{
mom(rnorm(100,0,1))

}
