% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{project}
\alias{project}
\title{Project Parameters back to the Feasible Space}
\usage{
project(theta, Theta)
}
\arguments{
\item{theta}{A vector of the parameters.}

\item{Theta}{A matrix each row of which have two scalars, 
first: lower bound, second: upper bound.}
}
\value{
A vector.
}
\description{
Project parameters back to the feasible space.
}
\examples{
theta = c(-0.3, 3, 0.4, 0.05)
Theta = matrix(c(0,10, 2,10, 0.1,0.9, 0.01,0.4), 
                 nrow=4, ncol=2, byrow=TRUE)
project(theta, Theta)

}
