% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{project2}
\alias{project2}
\title{More Fine Projection}
\usage{
project2(theta, Theta)
}
\arguments{
\item{theta}{A vector of the parameters.}

\item{Theta}{A matrix each row of which have two scalars, 
first: lower bound, second: upper bound.}
}
\value{
A vector.
}
\description{
Fine tune the mean and standard deviation of the normal distribution,
considering that it should not produce random variable smaller than 0,
or larger than 1.
}
\examples{
theta = c(-0.3, 3, 0.4, 1)
Theta = matrix(c(0,10, 2,10, 0.1,0.9, 0.01,0.4), 
                 nrow=4, ncol=2, byrow=TRUE)
project2(theta, Theta)

}
