% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{q}
\alias{q}
\title{Evaluate the Integrand Function}
\usage{
q(x, g, N, mu, sigma, w0 = 0.1)
}
\arguments{
\item{x}{A scalar variable.}

\item{g}{The parameter \eqn{g}.}

\item{N}{The parameter \eqn{N}, it should be larger than 2.}

\item{mu}{The parameter \eqn{\mu}, mean of the normal distribution.}

\item{sigma}{The parameter \eqn{\sigma}, standard deviation of the normal
distribution.}

\item{w0}{The parameter \eqn{w_0}, default to 0.1.}
}
\value{
A scalar value.
}
\description{
Compute the integrand function of the model which is described by the 
following equation:
\deqn{
  \beta(\alpha) = 1 - e^{\int_{-\infty}^{\alpha}q(x)dx},
}
where \eqn{\alpha \sim \mathcal{N}(\mu,\sigma^2)},
the integrand is
\deqn{
  q(x) \equiv \frac{g(N-1)F^{N-2}(x)f(x)x}{(w_0 + gF^{N-1}(x))(x-1)}.
}
and \eqn{f(x)} and \eqn{F(x)} are the PDF and CDF of the normal 
distribution \eqn{\mathcal{N}(\mu,\sigma^2)}, respectively.
}
\examples{
q(0.1, 0.2, 3.5, 0.5, 0.1)

}
