% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{q_beta}
\alias{q_beta}
\title{Evalute the Integrand Function}
\usage{
q_beta(x, g, N, a, b, w0 = 0.1)
}
\description{
Analog to [q()], but here the model is set with beta distribution.
}
\seealso{
[q()]
}
