% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sa}
\alias{sa}
\title{Stochastic Approximation}
\usage{
sa(theta0, Theta, stepsize, L, moments, G = 200L)
}
\arguments{
\item{theta0}{A vector of four scalars, i.e., the initial guess of the
parameters' value.}

\item{Theta}{A matrix of \eqn{4\times 2}, feasible space of the parameters.
Each row have two scalars, first: lower bound, second: upper bound.}

\item{stepsize}{A vector of four scalars, step size for each parameter.}

\item{L}{An integer, sample length used for each simulation embedded in the
SA searching.}

\item{moments}{A vector of the true sample moments, 1st moment,
2nd - 4th central moments.}

\item{G}{An integer, number of searching iterations.}
}
\value{
A matrix of the search trajectory, the first four columns
corresponding to the four parameters, and the last column corresponding
to the objective function value.
}
\description{
Search the optimal parameters by Stochastic Approximation (SA). That is
to find the parameters that minimize the objective (error) function.
}
\examples{
Y_sim = sim_Y(0.2, 3.5, 0.5, 0.1, 1000)    # 仿真生成1,000个模型样本
L = length(Y_sim)  
moments = mom(Y_sim)                       # 统计出一阶矩，2-4阶中心矩
lbub = c(0.1,5, 2.1,10, 0.1,0.9, 0.01,0.4) # 参数搜索范围
# 表示：0.1 <g< 5, 2.1 <N< 10, 0.1 <mu< 0.9, 0.01 <sigma< 0.4
Theta = matrix(lbub, nrow=4, ncol=2, byrow=TRUE) # 从数列转换成矩阵
# SA 搜索，限定搜索20步，步长固定为0.01（与1/k序列不一样）
theta0 = c(0.25, 3.0, 0.4, 0.05)           # 初始参数值 (g, N, mu, sigma)
stepsize = rep(0.01, 4)                    # 步长固定为0.01
theta_f_record = sa(theta0, Theta, stepsize, L, moments, G=20)

}
