% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sa_beta}
\alias{sa_beta}
\title{Stochastic Approximation}
\usage{
sa_beta(theta0, Theta, stepsize, L, moments, G = 200L)
}
\description{
Analog to [sa()], but here the model is set with beta distribution.
}
\seealso{
[sa()]
}
