% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_Y}
\alias{sim_Y}
\title{Simulate Samples of Variable Y}
\usage{
sim_Y(g, N, mu, sigma, L, w0 = 0.1)
}
\arguments{
\item{g}{The parameter \eqn{g}.}

\item{N}{The parameter \eqn{N}, it should be larger than 2.}

\item{mu}{The parameter \eqn{\mu}, mean of the normal distribution.}

\item{sigma}{The parameter \eqn{\sigma}, standard deviation of the normal
distribution.}

\item{L}{Length of the samples to generate.}

\item{w0}{The parameter \eqn{w_0}, default to 0.1.}
}
\value{
A vector of simulated \eqn{Y} samples.
}
\description{
Simulate samples of the variable
\deqn{
  Y \equiv \log(1 - \beta(\alpha)), Y = \int_{-\infty}^X q(x) dx.
}
where \eqn{\quad X \equiv \alpha}.
}
\examples{
sim_Y(0.2, 3.5, 0.5, 0.1, 100)

}
