% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_Y_beta}
\alias{sim_Y_beta}
\title{Simulate Samples of Variable Y}
\usage{
sim_Y_beta(g, N, a, b, L, w0 = 0.1)
}
\description{
Analog to [sim_Y()], but here the model is set with beta distribution.
}
\seealso{
[sim_Y()]
}
